2002-05-08  Roland McGrath  <roland@frob.com>
	* dir-lookup.c (diskfs_S_dir_lookup): int -> size_t
	* dir-readdir.c (diskfs_S_dir_readdir): Likewise.
	* file-get-trans.c (diskfs_S_file_get_translator): Likewise.
	* fsys-getroot.c (diskfs_S_fsys_getroot): Likewise.
	* diskfs.h (diskfs_append_args): unsigned -> size_t in argument type.
	(diskfs_append_std_options): Likewise.
	(diskfs_get_directs): Likewise.
	* file-exec.c (diskfs_S_file_exec): Likewise.
	* file-get-fs-opts.c (diskfs_S_file_get_fs_options): Likewise.
	* file-get-trans.c (diskfs_S_file_get_translator): Likewise.
	* file-getfh.c (diskfs_S_file_getfh): Likewise.
	* file-set-trans.c (diskfs_S_file_set_translator): Likewise.
	* fsys-getroot.c (diskfs_S_fsys_getroot): Likewise.
	* io-restrict-auth.c (diskfs_S_io_restrict_auth): Likewise.
	* opts-append-std.c (diskfs_append_std_options): Likewise.
	* opts-get.c (diskfs_append_args): Likewise.
	* dir-readdir.c (diskfs_S_dir_readdir): Likewise.

+++ libdiskfs/fsys-getroot.c
@@ -31,9 +31,9 @@ diskfs_S_fsys_getroot (fsys_t controlport,
 		       mach_msg_type_name_t replytype,
 		       mach_port_t dotdot,
 		       uid_t *uids,
-		       u_int nuids,
+		       size_t nuids,
 		       uid_t *gids,
-		       u_int ngids,
+		       size_t ngids,
 		       int flags,
 		       retry_type *retry,
 		       char *retryname,
@@ -101,7 +101,7 @@ diskfs_S_fsys_getroot (fsys_t controlport,
     {
       /* Handle symlink interpretation */
       char pathbuf[diskfs_root_node->dn_stat.st_size + 1];
-      int amt;
+      size_t amt;
 
       if (diskfs_read_symlink_hook)
 	error = (*diskfs_read_symlink_hook) (diskfs_root_node, pathbuf);
